Extract the CompilerHelpers module to spec/support
